Bootstrap shell template base
Checks for boolean fields that start with 'is_' for example:
is_active, is_confirmed, is_etc
shows an activate/deactivate button if is_active,
shows a yes/no label inestad of boolean 1/0

Checks for columns: filename, dir, mimetype and filesize
recognizes this columns and adds either the MeioUpload
or the CakeAttachment behavior to the model
(still needs some work in detecting several columns, works for 1 only)

Uses bootstrap buttons on the actions columns instead of just text.

On forms, it detects selects on add and sets the default value to empty
it also detects the file related columns and creates the form with
correct encrypt and adds the file type to the correct field.